Move refactor funccall to the method call
