+++ tests/Rector/MethodCall/RedirectBackHelperToBackHelperRector/Fixture/skip_with_arguments.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Laravel\Tests\Rector\FuncCall\HelperFuncCallToFacadeClassRector\Fixture;
+namespace Rector\Laravel\Tests\Rector\MethodCall\RedirectBackHelperToBackHelperRector\Fixture;
 
 class SkipWithStaticCall
 {
@@ -14,7 +14,7 @@ class SkipWithStaticCall
 -----
 <?php
 
-namespace Rector\Laravel\Tests\Rector\FuncCall\HelperFuncCallToFacadeClassRector\Fixture;
+namespace Rector\Laravel\Tests\Rector\MethodCall\RedirectBackHelperToBackHelperRector\Fixture;
 
 class SkipWithStaticCall
 {
